fixed one bug in BINFormatManager fixed savig ElasticContactParameters and SpheresContactGeometry, so that after loading the simulation is not going crazy added several try/catch blocks removed GLAPI and GLAPIENTRY from OpenGLWrapper XMLSaxParser is now better at counting lines
M    yade-doc/yade-deps.dia
M    yade-flat/src/src.pro
M    yade-guis/yade-gui-qt/src/QtGUI/YadeQtMainWindow.cpp
M    yade-core/src/yade/Interaction.cpp
M    yade-libs/yade-lib-serialization-xml/src/yade-lib-serialization-xml/XMLFormatManager.cpp
M    yade-libs/yade-lib-serialization-xml/src/yade-lib-serialization-xml/XMLSaxParser.cpp
M    yade-libs/yade-lib-serialization-bin/src/yade-lib-serialization-bin/BINFormatManager.cpp
M    yade-libs/yade-lib-opengl/src/yade-lib-opengl/OpenGLWrapper.hpp
M    yade-packages/yade-package-dem/src/DataClass/InteractionPhysics/ElasticContactParameters/ElasticContactParameters.cpp
M    yade-packages/yade-package-dem/src/DataClass/InteractionPhysics/ElasticContactParameters/ElasticContactParameters.hpp
M    yade-packages/yade-package-dem/src/DataClass/InteractionGeometry/SpheresContactGeometry/SpheresContactGeometry.cpp
M    yade-packages/yade-package-dem/src/DataClass/InteractionGeometry/SpheresContactGeometry/SpheresContactGeometry.hpp
D    yade-packages/yade-package-dem/src/Engine/EngineUnit/ElasticContactParameters
M    yade-packages/yade-package-dem/src/Engine/StandAloneEngine/ElasticCriterionTimeStepper/ElasticCriterionTimeStepper.cpp

+++ yade-libs/yade-lib-serialization-bin/src/yade-lib-serialization-bin/BINFormatManager.cpp
@@ -247,7 +247,7 @@ void BINFormatManager::serializeContainer(ostream& stream, Archive& ac , int dep
 void BINFormatManager::deserializeContainer(istream& stream, Archive& ac, const string& str)
 {
 //cerr << "+ deserializeContainer start\n"; int cccc=0;
-	unsigned short size;
+	unsigned int size;
 	stream.read(&((reinterpret_cast<char*>(&size))[0]),1);
 	stream.read(&((reinterpret_cast<char*>(&size))[1]),1);
 	stream.read(&((reinterpret_cast<char*>(&size))[2]),1);
